Move runner module into __init__
